data_ingestion/analysis_agent/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
yfinance>=0.2.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
python-dotenv>=0.19.0
pydantic>=1.8.0 